Split inventoryagent between interface and implementation
