+++ cmd/agent/subcommands/flare/command.go
@@ -34,7 +34,7 @@ import (
 	"github.com/DataDog/datadog-agent/comp/core/workloadmeta"
 	"github.com/DataDog/datadog-agent/comp/core/workloadmeta/collectors"
 	"github.com/DataDog/datadog-agent/comp/metadata/host/hostimpl"
-	"github.com/DataDog/datadog-agent/comp/metadata/inventoryagent"
+	"github.com/DataDog/datadog-agent/comp/metadata/inventoryagent/inventoryagentimpl"
 	"github.com/DataDog/datadog-agent/comp/metadata/inventoryhost/inventoryhostimpl"
 	"github.com/DataDog/datadog-agent/comp/metadata/resources/resourcesimpl"
 	"github.com/DataDog/datadog-agent/pkg/api/util"
@@ -112,7 +112,7 @@ func Commands(globalParams *command.GlobalParams) []*cobra.Command {
 				flare.Module(),
 				diagnosesendermanagerimpl.Module(),
 				// We need inventoryagent to fill the status page generated by the flare.
-				inventoryagent.Module(),
+				inventoryagentimpl.Module(),
 				hostimpl.Module(),
 				inventoryhostimpl.Module(),
 				resourcesimpl.Module(),
