+++ comp/metadata/bundle.go
@@ -9,7 +9,7 @@ package metadata
 
 import (
 	"github.com/DataDog/datadog-agent/comp/metadata/host/hostimpl"
-	"github.com/DataDog/datadog-agent/comp/metadata/inventoryagent"
+	"github.com/DataDog/datadog-agent/comp/metadata/inventoryagent/inventoryagentimpl"
 	"github.com/DataDog/datadog-agent/comp/metadata/inventorychecks/inventorychecksimpl"
 	"github.com/DataDog/datadog-agent/comp/metadata/inventoryhost/inventoryhostimpl"
 	"github.com/DataDog/datadog-agent/comp/metadata/resources/resourcesimpl"
@@ -25,7 +25,7 @@ func Bundle() fxutil.BundleOptions {
 		runnerimpl.Module(),
 		resourcesimpl.Module(),
 		hostimpl.Module(),
-		inventoryagent.Module(),
+		inventoryagentimpl.Module(),
 		inventoryhostimpl.Module(),
 		inventorychecksimpl.Module(),
 	)
