+++ comp/metadata/inventoryagent/component.go
@@ -6,11 +6,6 @@
 // Package inventoryagent implements a component to generate the 'datadog_agent' metadata payload for inventory.
 package inventoryagent
 
-import (
-	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
-	"go.uber.org/fx"
-)
-
 // team: agent-shared-components
 
 // Component is the component type.
@@ -22,28 +17,4 @@ type Component interface {
 	GetAsJSON() ([]byte, error)
 	// Get returns a copy of the agent metadata. Useful to be incorporated in the status page.
 	Get() map[string]interface{}
-}
-
-// Module defines the fx options for this component.
-func Module() fxutil.Module {
-	return fxutil.Component(
-		fx.Provide(newInventoryAgentProvider))
-}
-
-// Mock implements mock-specific methods for the inventoryagent component.
-type Mock interface {
-	Component
-}
-
-// MockModule defines the fx options for the mock component.
-//
-// Usage:
-//
-//	fxutil.Test[dependencies](
-//	   t,
-//	   inventoryagent.MockModule(),
-//	)
-func MockModule() fxutil.Module {
-	return fxutil.Component(
-		fx.Provide(newMock))
 }
