+++ comp/metadata/inventoryagent/inventoryagentimpl/inventoryagent.go
@@ -3,7 +3,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package inventoryagent
+// Package inventoryagentimpl implements a component to generate the 'datadog_agent' metadata payload for inventory.
+package inventoryagentimpl
 
 import (
 	"context"
@@ -20,16 +21,24 @@ import (
 	"github.com/DataDog/datadog-agent/comp/core/log"
 	"github.com/DataDog/datadog-agent/comp/core/sysprobeconfig"
 	"github.com/DataDog/datadog-agent/comp/metadata/internal/util"
+	iainterface "github.com/DataDog/datadog-agent/comp/metadata/inventoryagent"
 	"github.com/DataDog/datadog-agent/comp/metadata/runner/runnerimpl"
 	"github.com/DataDog/datadog-agent/pkg/serializer"
 	"github.com/DataDog/datadog-agent/pkg/serializer/marshaler"
 	"github.com/DataDog/datadog-agent/pkg/util/flavor"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 	"github.com/DataDog/datadog-agent/pkg/util/hostname"
 	"github.com/DataDog/datadog-agent/pkg/util/installinfo"
 	"github.com/DataDog/datadog-agent/pkg/util/optional"
 	"github.com/DataDog/datadog-agent/pkg/util/scrubber"
 	"github.com/DataDog/datadog-agent/pkg/version"
 )
+
+// Module defines the fx options for this component.
+func Module() fxutil.Module {
+	return fxutil.Component(
+		fx.Provide(newInventoryAgentProvider))
+}
 
 var (
 	// for testing
@@ -81,7 +90,7 @@ type dependencies struct {
 type provides struct {
 	fx.Out
 
-	Comp          Component
+	Comp          iainterface.Component
 	Provider      runnerimpl.Provider
 	FlareProvider flaretypes.Provider
 }
