+++ comp/metadata/inventoryagent/inventoryagentimpl/mock.go
@@ -3,11 +3,31 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package inventoryagent
+package inventoryagentimpl
+
+import (
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+	"go.uber.org/fx"
+
+	iainterface "github.com/DataDog/datadog-agent/comp/metadata/inventoryagent"
+)
+
+// MockModule defines the fx options for the mock component.
+//
+// Usage:
+//
+//	fxutil.Test[dependencies](
+//	   t,
+//	   inventoryagent.MockModule(),
+//	)
+func MockModule() fxutil.Module {
+	return fxutil.Component(
+		fx.Provide(newMock))
+}
 
 type inventoryagentMock struct{}
 
-func newMock() Component {
+func newMock() iainterface.Component {
 	return &inventoryagentMock{}
 }
